I now reconstruct the solution from junction points and line coordinates. This have to be done, because in raw cases the algorithm corrups the results. However, I was not able to find that bug so I decided to reconstruct the solution from the information which were not corrupted.

+++ src/main/java/de/biomedical_imaging/ij/steger/Offset.java
@@ -23,6 +23,6 @@
 package de.biomedical_imaging.ij.steger;
 
 public class Offset {
-	public long x, y;
+	public int x, y;
 	
 }
